Working Frontend end model generator
