+++ src/interfaces/database/TableInterfaces.ts
@@ -1,6 +1,6 @@
 // This file is generated by generateDatabase.ts
 import * as enums from "../../interfaces/enums";
-import { MultiLingualString } from "../application/MultiLangualString";
+import { MultiLingualString } from "../../utils/i18n/MultiLingualString";
 import { Debug_Data, Games_Settings, TimelineEntries_Changes } from "../domain";
 
 export interface DatasheetsModel {
