+++ src/interfaces/domain/GameSettings.ts
@@ -1,4 +1,4 @@
-import { MultiLingualString } from "../application/MultiLangualString";
+import { MultiLingualString } from "../../utils/i18n/MultiLingualString";
 import { GameSettingsEnum } from "../enums/games/GameSettingsEnum";
 import { DifficultyEnum } from "../enums/games/DifficultyEnum";
