+++ src/utils/Markdown.ts
@@ -1,4 +1,4 @@
-import { createMultiLingualString, MultiLingualString } from "../interfaces/application/MultiLangualString";
+import { createMultiLingualString, MultiLingualString } from "./i18n/MultiLingualString";
 
 export function createBlockList(items: string[]): MultiLingualString {
     return createMultiLingualString(items.map(item => `\`${item}\``).join());
